+++ examples/posix/src/my_platform.c
@@ -110,9 +110,22 @@ static void posix_on_controller_data(uni_hid_device_t* d, uni_controller_t* ctl)
 
             // Debugging
             // Axis ry: control rumble
-            if ((gp->buttons & BUTTON_A) && d->report_parser.set_rumble != NULL) {
-                d->report_parser.set_rumble(d, 128, 128);
+            if ((gp->buttons & BUTTON_A) && d->report_parser.set_dual_rumble != NULL) {
+                d->report_parser.set_dual_rumble(d, 1000, 100, 0, 255);
             }
+
+            if ((gp->buttons & BUTTON_B) && d->report_parser.set_dual_rumble != NULL) {
+                d->report_parser.set_dual_rumble(d, 0, 100, 255, 0);
+            }
+
+            if ((gp->buttons & BUTTON_X) && d->report_parser.set_dual_rumble != NULL) {
+                d->report_parser.set_dual_rumble(d, 0, 100, 255, 255);
+            }
+
+            if ((gp->buttons & BUTTON_Y) && d->report_parser.set_dual_rumble != NULL) {
+                d->report_parser.set_dual_rumble(d, 0, 100, 0, 255);
+            }
+
             // Buttons: Control LEDs On/Off
             if ((gp->buttons & BUTTON_B) && d->report_parser.set_player_leds != NULL) {
                 d->report_parser.set_player_leds(d, leds++ & 0x0f);
@@ -187,7 +200,11 @@ static posix_instance_t* get_posix_instance(uni_hid_device_t* d) {
 static void trigger_event_on_gamepad(uni_hid_device_t* d) {
     posix_instance_t* ins = get_posix_instance(d);
 
-    if (d->report_parser.set_rumble != NULL) {
+    if (d->report_parser.set_dual_rumble != NULL) {
+        d->report_parser.set_dual_rumble(
+            d, 0 /* delayed start ms */, 150 /* duration ms */, 0 /* weak_magnitude */, 255 /* strong_magnitude */
+        );
+    } else if (d->report_parser.set_rumble != NULL) {
         d->report_parser.set_rumble(d, 0x80 /* value */, 15 /* duration */);
     }