+++ src/components/bluepad32/parser/uni_hid_parser_stadia.c
@@ -18,15 +18,32 @@ struct stadia_ff_report {
     uint16_t weak_magnitude;
 } __attribute__((packed));
 
+enum {
+    STATE_RUMBLE_DISABLED,
+    STATE_RUMBLE_DELAYED,
+    STATE_RUMBLE_IN_PROGRESS,
+};
+
 // stadia_instance_t represents data used by the Stadia driver instance.
 typedef struct stadia_instance_s {
-    btstack_timer_source_t rumble_timer;
-    bool rumble_in_progress;
+    // Although technically, we can use one timer for delay and duration, easier to debug/maintain if we have two.
+    btstack_timer_source_t rumble_timer_duration;
+    btstack_timer_source_t rumble_timer_delayed_start;
+    int rumble_state;
+    // Used by delayed start
+    uint16_t rumble_weak_magnitude;
+    uint16_t rumble_strong_magnitude;
+    uint16_t rumble_duration_ms;
 } stadia_instance_t;
 _Static_assert(sizeof(stadia_instance_t) < HID_DEVICE_MAX_PARSER_DATA, "Stadia instance too big");
 
 static stadia_instance_t* get_stadia_instance(uni_hid_device_t* d);
+static void stadia_set_rumble_on(btstack_timer_source_t* ts);
 static void stadia_set_rumble_off(btstack_timer_source_t* ts);
+static void stadia_play_dual_rumble_now(struct uni_hid_device_s* d,
+                                        uint16_t duration_ms,
+                                        uint8_t weak_magnitude,
+                                        uint8_t strong_magnitude);
 
 void uni_hid_parser_stadia_setup(uni_hid_device_t* d) {
     if (d == NULL) {
@@ -41,23 +58,58 @@ void uni_hid_parser_stadia_setup(uni_hid_device_t* d) {
 }
 
 void uni_hid_parser_stadia_set_rumble(uni_hid_device_t* d, uint8_t value, uint8_t duration) {
-    uint8_t status;
+    uni_hid_parser_stadia_set_dual_rumble(d, 0, duration * 4, value, value);
+}
 
+void uni_hid_parser_stadia_set_dual_rumble(struct uni_hid_device_s* d,
+                                           uint16_t start_delay_ms,
+                                           uint16_t duration_ms,
+                                           uint8_t weak_magnitude,
+                                           uint8_t strong_magnitude) {
     if (d == NULL) {
         loge("Stadia: Invalid device\n");
         return;
     }
 
     stadia_instance_t* ins = get_stadia_instance(d);
-    if (ins->rumble_in_progress)
-        return;
+    switch (ins->rumble_state) {
+        case STATE_RUMBLE_DELAYED:
+            btstack_run_loop_remove_timer(&ins->rumble_timer_delayed_start);
+            break;
+        case STATE_RUMBLE_IN_PROGRESS:
+            btstack_run_loop_remove_timer(&ins->rumble_timer_duration);
+            break;
+        default:
+            // Do nothing
+    }
 
-    // TODO: It seems that the max value is 65356. Double check
-    uint16_t magnitude = value << 8;
+    if (start_delay_ms == 0) {
+        stadia_play_dual_rumble_now(d, duration_ms, weak_magnitude, strong_magnitude);
+    } else {
+        // Set timer to have a delayed start
+        ins->rumble_timer_delayed_start.process = &stadia_set_rumble_on;
+        ins->rumble_timer_delayed_start.context = d;
+        ins->rumble_state = STATE_RUMBLE_DELAYED;
+        ins->rumble_duration_ms = duration_ms;
+        ins->rumble_strong_magnitude = strong_magnitude;
+        ins->rumble_weak_magnitude = weak_magnitude;
+
+        btstack_run_loop_set_timer(&ins->rumble_timer_delayed_start, start_delay_ms);
+        btstack_run_loop_add_timer(&ins->rumble_timer_delayed_start);
+    }
+}
+
+static void stadia_play_dual_rumble_now(struct uni_hid_device_s* d,
+                                        uint16_t duration_ms,
+                                        uint8_t weak_magnitude,
+                                        uint8_t strong_magnitude) {
+    uint8_t status;
+
+    stadia_instance_t* ins = get_stadia_instance(d);
 
     const struct stadia_ff_report ff = {
-        .strong_magnitude = magnitude,
-        .weak_magnitude = magnitude,
+        .strong_magnitude = strong_magnitude << 8,
+        .weak_magnitude = weak_magnitude << 8,
     };
 
     status = hids_client_send_write_report(d->hids_cid, STADIA_RUMBLE_REPORT_ID, HID_REPORT_TYPE_OUTPUT,
@@ -68,12 +120,11 @@ void uni_hid_parser_stadia_set_rumble(uni_hid_device_t* d, uint8_t value, uint8_
     }
 
     // Set timer to turn off rumble
-    ins->rumble_timer.process = &stadia_set_rumble_off;
-    ins->rumble_timer.context = d;
-    ins->rumble_in_progress = 1;
-    int ms = duration * 4;  // duration: 256 ~= 1 second
-    btstack_run_loop_set_timer(&ins->rumble_timer, ms);
-    btstack_run_loop_add_timer(&ins->rumble_timer);
+    ins->rumble_timer_duration.process = &stadia_set_rumble_off;
+    ins->rumble_timer_duration.context = d;
+    ins->rumble_state = STATE_RUMBLE_IN_PROGRESS;
+    btstack_run_loop_set_timer(&ins->rumble_timer_duration, duration_ms);
+    btstack_run_loop_add_timer(&ins->rumble_timer_duration);
 }
 
 //
@@ -82,14 +133,22 @@ void uni_hid_parser_stadia_set_rumble(uni_hid_device_t* d, uint8_t value, uint8_
 static stadia_instance_t* get_stadia_instance(uni_hid_device_t* d) {
     return (stadia_instance_t*)&d->parser_data[0];
 }
+
+static void stadia_set_rumble_on(btstack_timer_source_t* ts) {
+    uni_hid_device_t* d = ts->context;
+    stadia_instance_t* ins = get_stadia_instance(d);
+
+    stadia_play_dual_rumble_now(d, ins->rumble_duration_ms, ins->rumble_weak_magnitude, ins->rumble_strong_magnitude);
+}
+
 static void stadia_set_rumble_off(btstack_timer_source_t* ts) {
     uint8_t status;
     uni_hid_device_t* d = ts->context;
     stadia_instance_t* ins = get_stadia_instance(d);
 
     // No need to protect it with a mutex since it runs in the same main thread
-    assert(ins->rumble_in_progress);
-    ins->rumble_in_progress = 0;
+    assert(ins->rumble_state != STATE_RUMBLE_DISABLED);
+    ins->rumble_state = STATE_RUMBLE_DISABLED;
 
     const struct stadia_ff_report ff = {
         .strong_magnitude = 0,